First push of a new system for the mapUpdateManager, spliting up most of the tasks and spliting the actual logic into the commands
