+++ src/main/java/net/countercraft/movecraft/mapUpdater/update/UpdateCommand.java
@@ -1,0 +1,5 @@
+package net.countercraft.movecraft.mapUpdater.update;
+
+public interface UpdateCommand {
+    public void doUpdate();
+}
